01 : Drop boot delay
The delay impacts other slaves and the timer which might
trigger other slaves watchdogs. Instead of applying an
expensive boot_delay, we increase the timer rate but drop
the boot delay. The restart pin is pretty sensitive, so,
we do not need to have larger power loss time just a quick
voltage on the pin

Signed-off-by: Hendrik Kahlen <hendrikkahlen.hk@gmail.com>